% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-get_question_column.R
\name{.get_question_column}
\alias{.get_question_column}
\title{Get the column number for the question}
\usage{
.get_question_column(wrp_question_uid)
}
\arguments{
\item{wrp_question_uid}{numeric reflecting which question has been selected}
}
\description{
Get the column number for the question
}
