% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-print_one_year_summary.R
\name{.print_one_year_summary}
\alias{.print_one_year_summary}
\title{Prints question summary for one year}
\usage{
.print_one_year_summary(question_summary)
}
\arguments{
\item{question_summary}{data.frame, summarises selected question}
}
\description{
Prints question summary for one year
}
