% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-print_question_summary.R
\name{.print_question_summary}
\alias{.print_question_summary}
\title{Wrapper function for question summaries}
\usage{
.print_question_summary(question_summary)
}
\arguments{
\item{question_summary}{data.frame, summarises selected question}
}
\description{
Wrapper function for question summaries
}
