% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-print_two_year_summary.R
\name{.print_two_year_summary}
\alias{.print_two_year_summary}
\title{Prints question summary for two years}
\usage{
.print_two_year_summary(question_summary)
}
\arguments{
\item{question_summary}{data.frame, summarises selected question}
}
\description{
Prints question summary for two years
}
