% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-wrp_cache.R
\name{.wrp_cache}
\alias{.wrp_cache}
\title{Download World Risk Poll data for this package if required.}
\usage{
.wrp_cache()
}
\description{
Ensure that the optional data is available locally in the
package cache. Will try to download the data only if it is not available.
}
