% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-wrp_folder_and_filepath.R
\name{.wrp_package_folder}
\alias{.wrp_package_folder}
\title{Package cache for World Risk Poll data}
\usage{
.wrp_package_folder()
}
\value{
folderpath to data
}
\description{
Caches the World Risk Poll data to your computer
}
