% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-wrp_processing.R
\name{.wrp_processing}
\alias{.wrp_processing}
\title{Processes World Risk Poll data for this package if required.}
\usage{
.wrp_processing()
}
\description{
Ensure that the optional data is available locally in the
package cache. Will try to download the data only if it is not available.
}
