% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-wrp_sysdata.R
\name{.wrp_sysdata_file_path}
\alias{.wrp_sysdata_file_path}
\title{Get aggregated data for a question in the World Risk Poll}
\usage{
.wrp_sysdata_file_path()
}
\value{
data frame with aggregated World Risk Poll question data
}
\description{
Allows you to access aggregated data for a World Risk Poll question.
}
\examples{
wrp_get(geography = "country", wrp_question_uid = "Q1")
}
