% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp.R
\docType{data}
\name{wrp}
\alias{wrp}
\title{World Risk Poll data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 280106 rows and 234 columns.
}
\source{
\url{https://wrp.lrfoundation.org.uk/}
}
\usage{
wrp
}
\description{
A dataset containing individual-level survey data from the World Risk Poll.
}
\keyword{datasets}
