% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_clean.R
\name{wrp_clean}
\alias{wrp_clean}
\title{Get aggregated data for a question in the World Risk Poll}
\usage{
wrp_clean(df)
}
\arguments{
\item{df}{data frame, a demographic category by which to aggregate}
}
\value{
data frame with aggregated World Risk Poll question data
}
\description{
Allows you to access aggregated data for a World Risk Poll question.
}
