% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_dictionary.R
\docType{data}
\name{wrp_dictionary}
\alias{wrp_dictionary}
\title{World Risk Poll data}
\format{
An object of class \code{look_for} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 234 rows and 9 columns.
}
\source{
\url{https://wrp.lrfoundation.org.uk/}
}
\usage{
wrp_dictionary
}
\description{
A codebook World Risk Poll.
}
\keyword{datasets}
