% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_get.R
\name{wrp_get}
\alias{wrp_get}
\title{Get aggregated data for a question in the World Risk Poll}
\usage{
wrp_get(geography = NULL, question = NULL)
}
\arguments{
\item{aggregation}{string, a demographic category by which to aggregate}

\item{survey_question}{string, the code for the survey question to focus on}

\item{tagged}{logical, if TRUE, retrieve human-readable response and aggregations, if FALSE, retrieve dbl+lab}
}
\value{
data frame with aggregated World Risk Poll question data
}
\description{
Allows you to access aggregated data for a World Risk Poll question.
}
