% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_visualise.R
\name{wrp_map}
\alias{wrp_map}
\title{Map the distribution of responses to chosen World Risk Poll questions}
\usage{
wrp_map(survey_question, survey_response)
}
\arguments{
\item{survey_question}{string, the code for the survey question to focus on}

\item{survey_response}{string, the selected reponse to the chosen survey question}
}
\value{
map
}
\description{
Provides a map of the proportion of people in each country who responded to a
selected World Risk Poll question.
}
