% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-wrp_remove.R
\name{wrp_remove}
\alias{wrp_remove}
\title{Delete all data in the package cache.}
\usage{
wrp_remove()
}
\value{
integer. The return value of the unlink() call: 0 for success,
1 for failure. See the unlink() documentation for details.
}
\description{
Delete all data in the package cache.
}
