% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_question_column.R
\name{.get_question_column}
\alias{.get_question_column}
\title{Get the column number for the question}
\usage{
.get_question_column(question_uid = wrp_questions$WRP_UID)
}
\arguments{
\item{question_uid}{numeric reflecting which question has been selected}
}
\description{
Get the column number for the question
}
