% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-get_regional_column.R
\name{.get_regional_column}
\alias{.get_regional_column}
\title{Get the column number for the geography selected}
\usage{
.get_regional_column(geography, wrp)
}
\arguments{
\item{geography}{string, a demographic category by which to aggregate, needs
to be one of "country", "region" or "income"}

\item{wrp, }{list, formatted World Risk Poll Questions}
}
\description{
Get the column number for the geography selected
}
