% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_regional_column.R
\name{.get_regional_column}
\alias{.get_regional_column}
\title{Get the column number for the geography selected}
\usage{
.get_regional_column(geography = wrp_regions$WRP_UID)
}
\arguments{
\item{geography}{string, a demographic category by which to aggregate, needs
to be one of "country", "region" or "income"}
}
\description{
Get the column number for the geography selected
}
