% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-get_weight_column.R
\name{.get_weight_column}
\alias{.get_weight_column}
\title{Get the column number for the weights}
\usage{
.get_weight_column(wrp_geography_col, wrp)
}
\arguments{
\item{wrp_geography_col}{numeric reflecting which regions have been selected}

\item{wrp, }{list, formatted World Risk Poll Questions}
}
\description{
Get the column number for the weights
}
