% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-wrp_check.R
\name{.wrp_check}
\alias{.wrp_check}
\title{GCheck World Risk Poll Data has previously been cached}
\usage{
.wrp_check()
}
\value{
data frame with aggregated World Risk Poll question data
}
\description{
Allows you to access aggregated data for a World Risk Poll question.
}
\examples{
wrp_get(geography = "country", wrp_question_uid = "Q1")
}
