% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-wrp_sysdata.R
\name{.wrp_sysdata_file_path}
\alias{.wrp_sysdata_file_path}
\title{Package cache for World Risk Poll data}
\usage{
.wrp_sysdata_file_path()
}
\value{
data frame with aggregated World Risk Poll question data
}
\description{
Caches the World Risk Poll data to your computer
}
\examples{
wrp_get(geography = "country", wrp_question_uid = "Q1")
}
