% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wrp}
\alias{wrp}
\title{World Risk Poll data}
\format{

}
\source{
\url{https://wrp.lrfoundation.org.uk/}
}
\usage{
wrp
}
\description{
A dataset containing individual-level survey data from the World Risk Poll.
}
\keyword{datasets}
