% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-wrp_aggregate.R
\name{wrp_aggregate}
\alias{wrp_aggregate}
\title{Get aggregated data for a question in the World Risk Poll}
\usage{
wrp_aggregate(df)
}
\arguments{
\item{df}{data frame, a demographic category by which to aggregate}
}
\value{
data frame with aggregated World Risk Poll question data
}
\description{
Allows you to access aggregated data for a World Risk Poll question.
}
