% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrp_get.R
\name{wrp_get}
\alias{wrp_get}
\title{Get aggregated data for a question in the World Risk Poll}
\usage{
wrp_get(geography = "country", wrp_question_uid = "Q1")
}
\arguments{
\item{geography}{string, a demographic category by which to aggregate,
needs to be one of "country", "region" or "income"}

\item{wrp_question_uid}{string, the number code for the survey
question to focus on}
}
\value{
data frame with aggregated World Risk Poll question data
}
\description{
Allows you to access aggregated data for a World Risk Poll question.
}
\examples{
wrp_get(geography = "country", wrp_question_uid = "Q1")
}
