% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-wrp_search.R
\name{wrp_search}
\alias{wrp_search}
\title{Get aggregated data for a question in the World Risk Poll}
\usage{
wrp_search(string = "violence")
}
\arguments{
\item{string}{string, search term for the World Risk Poll questions}
}
\value{
data frame with World Risk Poll questions containing the search term.
}
\description{
Allows you to access aggregated data for a World Risk Poll question.
}
\examples{
wrp_search(string = "violence")
}
